WIP: Change abstraction for loading of library.
Changing how to load libraries to ease reuse of functionality when we want to load two libraries. Currently loading of highlevel libraries are converted on windows, and will fail on linux and mac, though fixes are simple. The advantage of this way of loading is that all the specification of functions can be put into a single file rather than in platform specific folders, and that loading a different library do not need to duplicate the loader functions.