WIP: Add timing to return for all RTT functions.
This is a feature for applications that need information about some data was posted. This will never be accurate, but as long as the application read as fast as possible (i.e. calls itself in its own callback), this is as accurate you will be able to get a timestamp. The details of the timestamp are not fixed yet, and will most likely change.